fix(data-proxy): support paths without params
Also add a regression test. We have to spin up a new data proxy server
in a different port every test as for some reason calling
`await server.stop()` doesn't fully close the proxy server and the next
test will spin up a new server on the same port, but the old server
receives the request.