feat(data-proxy): conditionally retry proof verification
We've observed some problems where it seems that the RPC which the
overlay uses to query eligibility is ahead of the the RPC that the data
proxy uses to query eligibility. We've added extra log statements to
help debug this as well as retry logic when we have reason to believe
the RPC is out of sync.